requirements.txt
absl-py==0.9.0
asgiref==3.2.3
asn1crypto==1.2.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
blis==0.4.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
fastprogress==0.2.2
filelock==3.0.12
Flask==1.1.1
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
inflect==4.1.0
itsdangerous==1.1.0
jaraco.itertools==5.0.0
jedi==0.16.0
Jinja2==2.11.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
mkl
more-itertools==8.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numexpr==2.7.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.2.0
packaging==20.1
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
plyfile==0.7.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
sqlparse==0.3.1
srsly==1.0.1
tensorboardx
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
torchfile==0.1.0
tornado==6.0.3
traitlets==4.3.3
traittypes==0.2.1
trimesh==3.6.1
urllib3==1.24.2
visdom==0.1.8.9
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
zipp==3.1.0
